# [train] Speed up downloading of large checkpoints